Working-ish, if bare with jdbc and database store.
Can upload manga, list the manga in the database, view manga, navigate using keyboard or clicking on the image.

I have however rethought the storage model and a content repository would really be the better way to go.  So going to tear out the jdbc stuff and switch it for JCR (java content repository), specifically apache jackrabbit.