Added error login page, favicon.
Fixed a bug that overwrote the uploader and upload date of a manga.  Fixed a bug where scangroup wasn't pulled out of the chapter node.  Added css theme to login page and the edit page.  Changed bouncer to pull from the ContextRelativeRedirectURL in the session instead of just redirect to the web root.  Fixed a typo in the warning page.  Changed other controllers to use the bounce page instead of handling redirects themselves.  Expanded the user toolbar, added a logout option.